requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
alembic==1.13.3
python-multipart==0.0.9
aio-pika==9.4.3
Pillow==10.4.0

# Testing dependencies
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-mock==3.12.0
httpx==0.27.0
pytest-cov==5.0.0